Allow "/" to be used as the basePath
By allowing root (`"/"`) as a base path, developers can mount
applications anywhere and have Hop's routing work correctly.

This commit makes one change to the application code - it de-duplicates
slashes within the route we generate. Previously, with `"/"` as the
`basePath`, Hop would combine the `basePath` with the generated location
(e.g. `/deeply/nested`). By concatenating these values and passing the
result to `History.setPath`, the browser would see `//deeply/nested` and
think it was a request to that URI (inferring protocol based on the
prefixed `//`) and subsequently fail.